Separate Artifact and Fixture managers from hooks
