+++ src/test/worlds/PlaywrightWorld.ts
@@ -1,6 +1,5 @@
-// world.ts
 import { World } from '@cucumber/cucumber';
-import { FixtureManager } from '../../hooks/hooks';
+import FixtureManager from '../../hooks/FixtureManager';
 
 export default class PlaywrightWorld extends World {
   fixture: FixtureManager;
